Bug fix
1. edit, delete video url을 통해서라도 영상 주인만 들어갈 수 있도록 하기. 2. delete video할 때, splice, indexOf를 이용하여 user내에 있는 videos arr에서도 같이 지워주기

+++ src/controllers/videoController.js
@@ -20,19 +20,28 @@ export const watch = async(req,res) => {
 
 
 export const getEdit = async(req,res) => {
-    const id = req.params.id;
+    const {id} = req.params;
+    const {user:{_id}} = req.session;
     const video = await Video.findById(id);
     if(!video){
         return res.status(404).render("404", { pageTitle: "Video not found." });
     }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     return res.render("edit",{pageTitle:`Editing ${video.title}`, video});
 }
+
 export const postEdit = async (req, res) => {//post를 하면 mongoDB 내의 값을 변경해줌.
+    const {user:{_id}} = req.session;
     const {id} = req.params;
     const {title, description, hashtags } = req.body;
     const video = await Video.exists({_id:id});//video: DB에서 검색한 영상 object , Video: 우리가 만든 비디오 모델임.
     if(!video){
         return rres.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
     }
     await Video.findByIdAndUpdate(id, {
         title:title,
@@ -73,9 +82,20 @@ export const postUpload = async(req,res) => {
 
 export const deleteVideo = async(req,res) => {
     const { id } = req.params;
+    const {user:{_id}} = req.session;
+    const video = await Video.findById(id);
+    const user = await User.findById(_id);
+    if(!video){
+        return rres.status(404).render("404", { pageTitle: "Video not found." });
+    }
+    if(String(video.owner) !== String(_id)){
+        return res.status(403).redirect("/");
+    }
     await Video.findByIdAndDelete(id);
+    user.videos.splice(user.videos.indexOf(_id),1);
+    user.save();
     return res.redirect("/");
-}
+};
 
 export const search = async (req, res) => {
     //console.log(req.query);
